requirements.txt
python-dotenv
openai
vulture
aiohttp="3.9.0b0"